fix(daemon): drain transcription queue at shutdown
Final review found a three-task composition bug no per-task review could see. cancel fired at shutdown start, so transcribe_loop broke and dropped its receiver before audio_pipeline.stop() flushed the session's final partial segments. Those were then persisted and enqueued into a closed channel, so every clean exit dropped the tail un-transcribed, logged a false 'channel closed' WARN, and bumped transcription_dropped -- poisoning the one signal HEU-470 renders and HEU-473 tunes backlog from.

transcribe_loop no longer takes a cancellation token; it drains until the channel closes. main drops its sink Arc after audio_store_handle finishes, which is what closes the channel, and bounds the drain with a 10s timeout since an in-flight spawn_blocking whisper call cannot be aborted. The plan's 'await the transcription task last' mandate was vacuous -- the task had already exited -- so the stale comment is corrected too.

Also from review: guard the empty-transcript check with trim() (the Transcriber trait promises no trimming even though the engine impl trims); scale the granule clamp in u64 before narrowing to avoid overflow on a corrupt on-disk value; rewrite the transcription crate doc, which still claimed the crate only owned the path convention; mark update_transcript as the transcript-only variant with no production caller.

Replaces transcribe_loop_cancel_drops_queued with transcribe_loop_drains_all_queued_before_exit -- the old test asserted the behavior that caused the bug.

HEU-472.

+++ chronicle-daemon/crates/transcription/src/lib.rs
@@ -1,9 +1,17 @@
 //! Transcription pipeline for Chronicle.
 //!
-//! The whisper.cpp engine wiring lives in HEU-472. This crate currently
-//! owns the model-path convention and the allow-list of supported
-//! variants so HEU-472 (and the daemon startup check) agree on what is
-//! valid and where to look.
+//! This crate turns a persisted Ogg/Opus audio segment into text. It owns:
+//!
+//! - [`Transcriber`] — the trait the daemon depends on, so the pipeline can be
+//!   tested against a fake without a model on disk.
+//! - [`TranscriptionEngine`] — the whisper.cpp implementation (Metal by default,
+//!   `transcription-cpu` to opt out). Loads a ggml model once and shares it.
+//! - [`decode_opus_16k_mono`] — Opus → 16 kHz mono f32, the format whisper wants.
+//! - The model-path convention and the [`ModelVariant`] allow-list, so the daemon's
+//!   startup presence check and the loader agree on what is valid and where to look.
+//!
+//! The daemon drives all of this from `pipeline::transcribe_loop`; nothing here
+//! spawns tasks or touches the database.
 
 use std::fmt;
 use std::path::{Path, PathBuf};
@@ -186,7 +194,11 @@ pub fn decode_opus_16k_mono(path: &Path) -> Result<Vec<f32>, TranscriptionError>
     // left in by design (whisper tolerates it; over-trimming would be worse).
     // The clamp still protects against a future producer that writes a true
     // end-trimmed granule position.
-    let total16 = (last_granule as usize).saturating_sub(pre_skip) * DECODE_RATE / ENCODE_RATE;
+    // Scale in u64 before narrowing: `last_granule` comes off disk, and a corrupt
+    // value above ~1.15e15 would overflow the `* 16_000` if done in usize first
+    // (same class as the `n > buf.len()` decode guard above).
+    let total16 = (last_granule.saturating_sub(pre_skip as u64) * DECODE_RATE as u64
+        / ENCODE_RATE as u64) as usize;
     if total16 > 0 && pcm.len() > total16 {
         pcm.truncate(total16);
     }
